pyproject.toml
[project]
name = "documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "isort>=8.0.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
